defconfig: allow to use perf archive
When performing the system wide profiling/tracing lab, we analyze the
perf.data file on host, which then need some options to properly convert
addresses to symbols (-k, --symfs). An alternative to this is to use
perf archive, which is able to gather some build ids (and possibly
symbol table for kernel). perf archive calls a bash script on the target
which in turn runs perf buildid-list and create an archive with the
result. This commands then need bash, tar and bzip2.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>