package/pahole: fix build with recent host-cmake
With cmake >= 4.0.0, the debugging training build can fail because of a
new check on the minimal version requested by cmake projects.
host-pahole currently fails on this check. The upstream pahole project
has already fixed the issue in version 1.28. A patch has been sent to
upstream buildroot to fix the issue as well. As the 2024.02 branch is
not supported anymore , backporting the upstream patch is fine. It
should be dropped on the next buildroot update.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>