package/m4: fix build failure with gcc 15
This is a temporary workaround to address build issues with host-m4
when using GCC 15, as encountered on
the recently released Fedora 42. The issue stems from stricter
compiler behavior introduced in recent
bersions of GCC.

A proper fix has already been committed upstream and will be included
in the next release of GNU m4:
https://git.savannah.gnu.org/cgit/m4.git/commit/?h=branch-1.4&id=a22c9802dd7e724eaefb21dc21d84ac2d3a49c89

Until that version is available in buildroot, this patch ensures
continued compatibility and successful
builds.

More context:
https://lists.buildroot.org/pipermail/buildroot/2025-April/776922.html

Signed-off-by: Joseph Zikusooka (ZIK) <zik@jambula.net>

(The Buildroot upstream fix is based on BR2_HOST_GCC_AT_LEAST_15 which did
not exist in 2024.04, so applied the simple patch from:
https://lore.kernel.org/buildroot/20250503114339.13240-1-zik@jambula.net/)