defconfig: disable bash
bash was enabled by commit cdb4b6e72c30 ("defconfig: allow to use perf
archive") in order to enable 'perf archive' on the target. However that
never worked well and was never added to the training instructions. Now
bash does not build on systems with gcc 15, thus jsut remove it.

Suggested-by: Alexis Lothoré <alexis.lothore@bootlin.com>
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>